fibonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,7 +9,7 @@ int main(void)
 {
 	int i = 0;
 	long j = 1;
-	int k = 2;
+	long k = 2;
 	while (i < 50)
 	{
 		if (i == 0)
